MemSQL, Row/Char index menu, bug fixes and User agent
- New database vendor: MemSQL (in progress)
- New Option to disable injection of database's Metadata
- Tree Node Menu: Inject from/to no. of row/character
- Tree Node Menu: Dump to a file
- Colorize the timestamp in Console tab
- User agent configuration (in progress)
- Translation: Fix #45431 ru, Fix #44483 zh, Fix #45557 in_ID, Fix
#45420 fr
- Fix #45611: IndexOutOfBoundsException on getBoundsAt()
- Fix #45540: NullPointerException on node.getUserObject()
- Fix #45349: JSONException on new JSONObject(json)
- Fix #45348: IllegalComponentStateException on show()
- Fix #44642: NoClassDefFoundError on get()
- Fix #44641: ExceptionInInitializerError on get()
- Fix #44635: ArrayIndexOutOfBoundsException on getItem()
- Fix #43974: PatternSyntaxException on regexFilter() => Pattern.quote()
- Fix #43220: EmptyStackException on setText()
- Fix #42832: ClassCastException on showMessageDialog()
- Fix #42831: ClassCastException on showOpenDialog()

+++ src/main/java/com/jsql/view/swing/scrollpane/JScrollIndicator.java
@@ -266,7 +266,7 @@ public class JScrollIndicator extends JLayeredPane {
                 return color;
             }
             int rgb = color.getRGB() & 0xFFFFFF; // color without alpha values
-            rgb |= (int)(this.alpha/100*255) << 24; // add alpha value
+            rgb |= (this.alpha / 100 * 255) << 24; // add alpha value
             return new Color(rgb, true);
         }
